pyproject.toml
[tool.poetry]
name = "rdeciderbot"
version = "1.0.0"
description = """
Tired of having trouble deciding what to do with your friends?
Like having a list of series or movies and not being able to decide which one to watch
Or have a list of video games and can't decide which one to play
I'm here to help you choose a totally random and unbiased option to avoid that bullshit
"""
authors = ["Gabriel Anzola <ga.anzola15@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
python-telegram-bot = "^20.1"
python-dotenv = "^1.0.0"
pre-commit = "^3.1.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
